# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StealADeal Backend API
  version: 0.1.0
  description: |
    OpenAPI contract for the current H2-backed development backend.
    This spec reflects the implemented dealer, vehicle, lead, appointment,
    and dashboard endpoints.
servers:
  - url: http://localhost:8282
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/dealers:
    get:
      tags: [Dealers]
      summary: List dealers
      operationId: listDealers
      security: []
      responses:
        '200':
          description: Dealers returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dealer'
    post:
      tags: [Dealers]
      summary: Create dealer
      operationId: createDealer
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDealerRequest'
      responses:
        '201':
          description: Dealer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealer'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/dealers/{dealerId}:
    get:
      tags: [Dealers]
      summary: Get dealer by id
      operationId: getDealer
      security: []
      parameters:
        - $ref: '#/components/parameters/DealerId'
      responses:
        '200':
          description: Dealer returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealer'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Dealers]
      summary: Update dealer details
      operationId: updateDealer
      parameters:
        - $ref: '#/components/parameters/DealerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDealerRequest'
      responses:
        '200':
          description: Dealer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealerId}/approval:
    patch:
      tags: [Dealers]
      summary: Update dealer approval status
      operationId: updateDealerApproval
      parameters:
        - $ref: '#/components/parameters/DealerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDealerApprovalRequest'
      responses:
        '200':
          description: Dealer approval updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealer'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/deals/{dealId}:
    get:
      tags: [Deals]
      summary: Get deal by id
      operationId: getDeal
      parameters:
        - $ref: '#/components/parameters/DealId'
      responses:
        '200':
          description: Deal returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deal'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/inbox/dealers/{dealerId}:
    get:
      tags: [Inbox]
      summary: Get dealer inbox
      operationId: getDealerInbox
      parameters:
        - $ref: '#/components/parameters/DealerId'
      responses:
        '200':
          description: Dealer inbox returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealerInbox'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
  parameters:
    DealerId:
      in: path
      name: dealerId
      required: true
      schema:
        type: integer
        format: int64
...
    DealId:
      in: path
      name: dealId
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Dealer:
      type: object
      required: [id, name, licenseNumber, city, state, approved]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        licenseNumber:
          type: string
        city:
          type: string
        state:
          type: string
        approved:
          type: boolean
    CreateDealerRequest:
      type: object
      required: [name, licenseNumber, city, state]
      properties:
        name:
          type: string
        licenseNumber:
          type: string
        city:
          type: string
        state:
          type: string
          pattern: '^[A-Za-z]{2}$'
    UpdateDealerRequest:
      allOf:
        - $ref: '#/components/schemas/CreateDealerRequest'
    UpdateDealerApprovalRequest:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean
...
    Deal:
      type: object
      required:
        - id
        - vehicleId
        - vehicleTitle
        - buyerName
        - buyerEmail
        - buyerPhone
        - buyerAddressLine1
        - buyerCity
        - buyerState
        - buyerPostalCode
        - fulfillmentType
        - tradeIn
        - tradeInOffer
        - vehiclePrice
        - taxAmount
        - registrationFee
        - documentationFee
        - deliveryFee
        - discountAmount
        - depositAmount
        - depositPaid
        - totalAmount
        - stage
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
        vehicleId:
          type: integer
          format: int64
        vehicleTitle:
          type: string
        buyerName:
          type: string
        buyerEmail:
          type: string
          format: email
        buyerPhone:
          type: string
        buyerAddressLine1:
          type: string
        buyerAddressLine2:
          type: string
          nullable: true
        buyerCity:
          type: string
        buyerState:
          type: string
        buyerPostalCode:
          type: string
        fulfillmentType:
          $ref: '#/components/schemas/FulfillmentType'
        fulfillmentStatus:
          $ref: '#/components/schemas/FulfillmentStatus'
        fulfillmentScheduledAt:
          type: string
          format: date-time
          nullable: true
        fulfillmentLocation:
          type: string
          nullable: true
        fulfillmentNotes:
          type: string
          nullable: true
        tradeIn:
          type: boolean
        tradeInVin:
          type: string
          nullable: true
        tradeInMileage:
          type: integer
          nullable: true
        tradeInOffer:
          type: number
          format: double
        vehiclePrice:
          type: number
          format: double
        taxAmount:
          type: number
          format: double
        registrationFee:
          type: number
          format: double
        documentationFee:
          type: number
          format: double
        deliveryFee:
          type: number
          format: double
        discountAmount:
          type: number
          format: double
        depositAmount:
          type: number
          format: double
        depositPaid:
          type: boolean
        totalAmount:
          type: number
          format: double
        stage:
          $ref: '#/components/schemas/DealStage'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    InboxSummary:
      type: object
      required: [totalDeals, activeDeals, completedDeals, openTaskCount, inProgressTaskCount, unreadNotificationCount, readyForHandoffCount]
      properties:
        totalDeals:
          type: integer
        activeDeals:
          type: integer
        completedDeals:
          type: integer
        openTaskCount:
          type: integer
        inProgressTaskCount:
          type: integer
        unreadNotificationCount:
          type: integer
        readyForHandoffCount:
          type: integer
    QueueSummary:
      type: object
      required: [awaitingBuyerCount, needsDocumentReviewCount, readyForHandoffCount, stalledCount]
      properties:
        awaitingBuyerCount:
          type: integer
        needsDocumentReviewCount:
          type: integer
        readyForHandoffCount:
          type: integer
        stalledCount:
          type: integer
...
    InboxDeal:
      type: object
      required:
        [dealId, vehicleId, dealerId, vehicleTitle, stage, depositPaid, readyForHandoff, readyForCompletion,
         blockers, openTaskCount, unreadNotificationCount, nextAction, updatedAt]
      properties:
        dealId:
          type: integer
          format: int64
        vehicleId:
          type: integer
          format: int64
        dealerId:
          type: integer
          format: int64
        vehicleTitle:
          type: string
        stage:
          $ref: '#/components/schemas/DealStage'
        depositPaid:
          type: boolean
        fulfillmentScheduledAt:
          type: string
          format: date-time
          nullable: true
        fulfillmentLocation:
          type: string
          nullable: true
        readyForHandoff:
          type: boolean
        readyForCompletion:
          type: boolean
        blockers:
          type: array
          items:
            type: string
        openTaskCount:
          type: integer
        unreadNotificationCount:
          type: integer
        nextAction:
          type: string
        updatedAt:
          type: string
          format: date-time
    ParticipantInbox:
      type: object
      required: [participantType, participantReference, summary, deals, tasks, notifications]
      properties:
        participantType:
          $ref: '#/components/schemas/ParticipantType'
        participantReference:
          type: string
        summary:
          $ref: '#/components/schemas/InboxSummary'
        deals:
          type: array
          items:
            $ref: '#/components/schemas/InboxDeal'
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/DealTask'
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
    DealerInbox:
      type: object
      required: [inbox, queueSummary]
      properties:
        inbox:
          $ref: '#/components/schemas/ParticipantInbox'
        queueSummary:
          $ref: '#/components/schemas/QueueSummary'
...
    DealTask:
      type: object
      required: [id, dealId, code, assigneeType, assigneeReference, title, description, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        dealId:
          type: integer
          format: int64
        code:
          type: string
        assigneeType:
          $ref: '#/components/schemas/ParticipantType'
        assigneeReference:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/DealTaskStatus'
        dueAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Notification:
      type: object
      required: [id, recipientType, recipientReference, title, message, read, createdAt]
      properties:
        id:
          type: integer
          format: int64
        dealId:
          type: integer
          format: int64
          nullable: true
        recipientType:
          $ref: '#/components/schemas/ParticipantType'
        recipientReference:
          type: string
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
...
    FulfillmentType:
      type: string
      enum:
        - PICKUP
        - HOME_DELIVERY
...
    DealStage:
      type: string
      enum:
        - INITIATED
        - OFFER_SENT
        - BUYER_CONFIRMED
        - DEPOSIT_PAID
        - DOCUMENTS_PENDING
        - READY_FOR_HANDOFF
        - COMPLETED
        - CANCELED
...
    ParticipantType:
      type: string
      enum:
        - BUYER
        - DEALER
        - ADMIN
    DealTaskStatus:
      type: string
      enum:
        - OPEN
        - IN_PROGRESS
        - COMPLETED
        - CANCELED
    FulfillmentStatus:
      type: string
      enum:
        - UNSCHEDULED
        - SCHEDULED
        - READY
        - COMPLETED
        - CANCELED
...